backend/requirements.txt
# Web API
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0

# LLM and Vector Database
langchain==0.0.334
langchain-google-genai==0.0.5
google-generativeai==0.3.2
chromadb==0.4.17

# Document processing
langchain-text-splitters==0.0.1
beautifulsoup4==4.12.2
python-docx==0.8.11
PyPDF2==3.0.1
markdown==3.5.1

# Embeddings
sentence-transformers==2.2.2

# Testing
pytest==7.4.3
httpx==0.25.2

# Utilities
pandas==2.1.3
numpy==1.26.2
tqdm==4.66.1